+++ migrations/04_20230706210246_supervisor_id_table.js
@@ -4,7 +4,7 @@
  */
 exports.up = function(knex) {
   return knex.schema.alterTable('users', table => {
-    table.integer('supervisor_id');
+    table.bigint('supervisor_id');
     table.foreign('supervisor_id').references('users.id').
     deferrable('deferred')
   })
